backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "FastAPI Chatbot Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "python-socketio>=5.11.0",
    "python-jose[cryptography]>=3.3.0",

    # IMPORTANT:
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",    # ← FIXES YOUR ERROR

    "python-multipart>=0.0.6",
    "pydantic[email]>=2.0.0",
]
